refactor: common-service (#787)
* feat: user-service

* refactor: common-service

* chore: upgrade version

* chore: update version

* chore: version and changelog

* chore: update changelog
