+++ extensions/iceworks-config-helper/package.json
@@ -3,7 +3,7 @@
   "displayName": "Config Helper",
   "description": "Easily write Config files in icejs(& rax-app)",
   "publisher": "iceworks-team",
-  "version": "0.5.4",
+  "version": "0.5.5",
   "engines": {
     "vscode": "^1.41.0"
   },
@@ -121,9 +121,9 @@
     "webpack-cli": "^3.3.12"
   },
   "dependencies": {
-    "@iceworks/common-service": "^0.1.14",
+    "@iceworks/common-service": "^0.2.0",
     "@iceworks/i18n": "^0.1.0",
-    "@iceworks/project-service": "^0.1.15",
+    "@iceworks/project-service": "^0.2.0",
     "@iceworks/recorder": "^0.1.6",
     "@iceworks/vscode-webview": "^0.1.5",
     "generate-schema": "^2.6.0",
