+++ extensions/iceworks-material-helper/package.json
@@ -375,10 +375,10 @@
     "@babel/parser": "^7.10.2",
     "@babel/traverse": "^7.10.1",
     "@babel/types": "^7.10.2",
-    "@iceworks/common-service": "^0.1.3",
+    "@iceworks/common-service": "^0.2.0",
     "@iceworks/i18n": "^0.1.0",
-    "@iceworks/material-engine": "^0.1.0",
-    "@iceworks/project-service": "^0.1.28",
+    "@iceworks/material-engine": "^0.2.0",
+    "@iceworks/project-service": "^0.2.0",
     "@iceworks/project-utils": "^0.3.0",
     "@iceworks/recorder": "^0.1.6",
     "@iceworks/vscode-webview": "^0.1.1",
