+++ extensions/iceworks-project-creator/package.json
@@ -3,7 +3,7 @@
   "displayName": "Application Creator",
   "description": "Quick create a Universal Application(React/Rax/Vue, etc).",
   "publisher": "iceworks-team",
-  "version": "0.4.9",
+  "version": "0.4.10",
   "engines": {
     "vscode": "^1.41.0"
   },
@@ -66,11 +66,11 @@
     "webpack-cli": "^3.3.11"
   },
   "dependencies": {
-    "@iceworks/common-service": "^0.1.3",
+    "@iceworks/common-service": "^0.2.0",
     "@iceworks/i18n": "^0.1.0",
-    "@iceworks/material-engine": "^0.1.0",
+    "@iceworks/material-engine": "^0.2.0",
     "@iceworks/pegasus-service": "^0.1.2",
-    "@iceworks/project-service": "^0.1.4",
+    "@iceworks/project-service": "^0.2.0",
     "@iceworks/recorder": "^0.1.6",
     "@iceworks/vscode-webview": "^0.1.1",
     "js-base64": "^3.6.0"
