+++ extensions/iceworks-project-creator/src/services/index.ts
@@ -3,5 +3,6 @@ import * as material from '@iceworks/material-engine/lib/material';
 import * as scaffold from '@iceworks/material-engine/lib/scaffold';
 import * as common from '@iceworks/common-service';
 import * as pegasus from '@iceworks/pegasus-service';
+import * as user from '@iceworks/user-service';
 
-export default { project, common, material, pegasus, scaffold };
+export default { project, common, material, pegasus, scaffold, user };
