+++ extensions/iceworks-time-master/package.json
@@ -3,7 +3,7 @@
   "displayName": "Time Master",
   "description": "Metrics, insights, and time tracking automatically generated from your programming activity.",
   "publisher": "iceworks-team",
-  "version": "0.4.2",
+  "version": "0.4.3",
   "preview": true,
   "engines": {
     "vscode": "^1.41.0"
@@ -91,7 +91,7 @@
     "webpack-cli": "^3.3.11"
   },
   "dependencies": {
-    "@iceworks/common-service": "^0.1.0",
+    "@iceworks/common-service": "^0.2.0",
     "@iceworks/project-utils": "^0.3.0",
     "@iceworks/constant": "^0.1.2",
     "@iceworks/i18n": "^0.1.0",
