+++ packages/common-service/src/template.ts
@@ -1,0 +1,7 @@
+export function getImportTemplate(name: string, source: string): string {
+  return `import ${name} from '${source}';\n`;
+}
+
+export function getTagTemplate(name: string): string {
+  return `<${name} /> \n`;
+}
