+++ packages/material-engine/CHANGELOG.md
@@ -1,0 +1,5 @@
+# CHANGELOG
+
+## 0.2.0
+
+- chore: upgrade @iceworks/common-service version
