+++ packages/material-engine/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/material-engine",
-  "version": "0.1.7",
+  "version": "0.2.0",
   "description": "Iceworks Material Engine for VSCode extension.",
   "files": [
     "lib"
@@ -13,10 +13,10 @@
     "@babel/traverse": "^7.11.5",
     "@babel/types": "^7.11.5",
     "@iceworks/code-generator": "^0.1.3",
-    "@iceworks/common-service": "^0.1.3",
+    "@iceworks/common-service": "^0.2.0",
     "@iceworks/i18n": "^0.1.0",
     "@iceworks/material-utils": "^0.1.3",
-    "@iceworks/project-service": "^0.1.4",
+    "@iceworks/project-service": "^0.2.0",
     "axios": "^0.20.0",
     "bluebird": "^3.7.2",
     "ejs": "^3.1.5",
