+++ packages/project-service/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/project-service",
-  "version": "0.1.28",
+  "version": "0.2.0",
   "description": "Iceworks project service for VSCode extension.",
   "files": [
     "lib"
@@ -8,7 +8,7 @@
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "dependencies": {
-    "@iceworks/common-service": "^0.1.3",
+    "@iceworks/common-service": "^0.2.0",
     "@iceworks/generate-project": "^1.0.7",
     "@iceworks/material-utils": "^0.1.3",
     "@iceworks/project-utils": "^0.3.0",
