+++ packages/user-service/src/types.ts
@@ -1,0 +1,5 @@
+export interface IUserInfo {
+  empId: string;
+  account: string;
+  gitlabToken: string;
+}
