+++ packages/vscode-generator/src/templates/extension/package.json.ejs
@@ -42,8 +42,8 @@
     ]
   },
   "dependencies": {
-    "@iceworks/common-service": "^0.1.14",
-    "@iceworks/project-service": "^0.1.17",
+    "@iceworks/common-service": "^0.2.0",
+    "@iceworks/project-service": "^0.2.0",
     "@iceworks/vscode-webview": "^0.1.5"
   },
   "devDependencies": {
